Fix silent failures, type mismatches, validation gaps, and test isolation (#145, #146, #147, #148, #149, #150, #151, #152, #153, #154, #155, #156)
- #145: Replace unwrap_or_default() with warn! logging in build_tracker and
  PolicyEngine when repo_owner/repo_name are missing.
- #146: Abort with ProviderError when single-PR mode has no
  comment_issue_number instead of fabricating a loop/0- branch.
- #147: Add hook_failed field to SessionSummary; on_complete hook non-zero
  exit now causes process exit code 1. Add eprintln alongside warn/error.
- #148: Apply redact_secrets() to gh pr merge stderr output.
- #149: Add warn! and eprintln when resolve_workspace_dir falls back to ".".
- #150: Add eprintln when Ctrl+C handler installation fails.
- #151: Replace silent .flatten() and Err(_) => return in prune_old_runs
  with warn! logging for read_dir failure and skipped entries.
- #152: Replace let _ suppression in service shutdown ACK with match that
  logs serialization and write errors.
- #153: Update Deref comment to WARNING; remove test that accessed raw
  fields through ValidatedLoopConfig Deref.
- #154: Add RulesConfig::validate() method; narrow resolve_rule_paths to
  accept &mut RulesConfig instead of &mut LoopConfig.
- #155: Add RulesCacheGuard RAII type for test isolation; replace manual
  clear_rules_cache() call pairs with guard in all tests.
- #156: Change derive_branch_name, BranchManager::branch_name,
  BranchManager::ensure_branch, and PrManager::handle_milestone from u64
  to u32 issue numbers, removing all as u64 casts.

+++ src/branch.rs
@@ -37,7 +37,7 @@ const DEFAULT_MAX_SLUG_LEN: usize = 40;
 /// ```
 pub fn derive_branch_name(
     prefix: &str,
-    issue_number: u64,
+    issue_number: u32,
     title: &str,
     max_slug_length: usize,
 ) -> String {
@@ -297,7 +297,7 @@ impl BranchManager {
     }
 
     /// Derive the feature branch name for this issue.
-    pub fn branch_name(&self, issue_number: u64, title: &str) -> String {
+    pub fn branch_name(&self, issue_number: u32, title: &str) -> String {
         derive_branch_name(
             &self.config.branch_prefix,
             issue_number,
@@ -311,7 +311,7 @@ impl BranchManager {
     /// reuses it without creating a duplicate.
     ///
     /// Returns the branch name.
-    pub fn ensure_branch(&self, issue_number: u64, title: &str) -> Result<String, BranchError> {
+    pub fn ensure_branch(&self, issue_number: u32, title: &str) -> Result<String, BranchError> {
         let branch = self.branch_name(issue_number, title);
 
         // Guard: never operate on base_branch
